Moved doc_test to message_test
