+++ waku/message_test.go
@@ -22,6 +22,8 @@ import (
 	"bytes"
 	"crypto/aes"
 	"crypto/cipher"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/stretchr/testify/require"
 	mrand "math/rand"
 	"testing"
 	"time"
@@ -494,4 +496,16 @@ func TestValidateAndParseSizeOfPayloadSize(t *testing.T) {
 			msg.ValidateAndParse()
 		})
 	}
+}
+
+func TestEncodeDecodeVersionedResponse(t *testing.T) {
+	response := NewMessagesResponse(common.Hash{1}, []EnvelopeError{{Code: 1}})
+	bytes, err := rlp.EncodeToBytes(response)
+	require.NoError(t, err)
+
+	var mresponse MultiVersionResponse
+	require.NoError(t, rlp.DecodeBytes(bytes, &mresponse))
+	v1resp, err := mresponse.DecodeResponse1()
+	require.NoError(t, err)
+	require.Equal(t, response.Response.Hash, v1resp.Hash)
 }
